+++ core/src/test/kotlin/sift/core/api/testdata/set1/ClassWithVisibility.kt
@@ -1,0 +1,12 @@
+package sift.core.api.testdata.set1
+
+class ClassWithVisibilityA {
+}
+
+internal class ClassWithVisibilityB {
+    internal fun fnInternal() = Unit
+    private fun fnPrivate() = Unit
+    protected fun fnProtected() = Unit
+    fun fnPublic() = Unit
+}
+